requirements.txt
selenium
pandas
python-dotenv
beautifulsoup4
openpyxl
flask
pytest
